Implement -split-sections in the NCG (for x86), #8405
This adds a flag -split-sections that does similar things to
-split-objs, but using sections in single object files instead of
relying on the Satanic Splitter and other abominations.

The --gc-sections flag (which allows unused sections to actually be
removed) is currently added unconditionally to all link commands, so
that space savings from having base compiled with sections can be
realized.

Not everything is roses though - if ghc itself is compiled with
sections, there are so many of them that linking plugins statically can
take up towards 10 minutes. (Using the gold linker reduces that time to
15 seconds.) Anyway, that motivates the use of a flag to control it,
and it's disabled for the compiler same as SplitObjs.